Move Disk to separate file
